Fixed: Moved processmon externally consumable API to processmon package
